+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added validation to ensure no roles or bindings have cluster-admin permissions.
+
 ## [0.41.0] - 2024-01-25
 
 ### Added